split up app and server in preperation to add testing
