+++ server/src/server.ts
@@ -1,8 +1,15 @@
 import express from "express";
 import path from "path";
 import * as dotenv from "dotenv";
-import routes from "./routes/index";
 import mongoose from "mongoose";
+import configureApp from "./app";
+
+// this is the server's main entrypoint from the beginning
+// It is responsible for setting up the entire app environment
+// as it was launched, including mongo setup and calling .listen
+// app express instance.  However, the app instance is setup
+// for handling its routes seperately in app.ts, so it may be
+// tested
 
 let production = process.argv.find(el => el == "--production");
 
@@ -12,29 +19,14 @@ if(production) {
     console.log("We're NOT running in production");
 }
 
-// I may need to do something different for production
 let envPath = production ? path.join(__dirname, "..", ".env") : path.join(__dirname, "..", "..", ".env");
 
 dotenv.config({path: envPath});
 
-const app:express.Application = express();
+const app:express.Application = configureApp(!!production);
 const port = process.env.PORT || 3001;
 
-app.use(express.urlencoded({ extended: true }));
-app.use(express.json());
-app.use(routes);
-
 mongoose.connect(process.env.MONGODB_URI || "mongodb://localhost/populate", { useNewUrlParser: true });
-
-let clientstaticpath = production ? path.join(__dirname, "clientbuild") : "";
-
-if(clientstaticpath != "") {
-    app.use(express.static(clientstaticpath));
-}
-
-let nonReactPublicsPath = path.join(__dirname,"nonReact_publics");
-
-app.use("/nonReact",express.static(nonReactPublicsPath));
 
 // start the Express server
 app.listen( port, () => {
